+++ src/components/auth/Registration/ExpertRegistrationFormContent.jsx
@@ -1,21 +1,21 @@
-import ExpertLoginForm from "./ExpertLoginForm";
-import LoginImage from "../LoginImage";
+import ExpertRegistrationForm from "./ExpertRegistrationForm";
+import LoginImage from "../Login/LoginImage";
 
-const ExpertLoginFormContent = () => {
+const ExpertRegistrationFormContent = () => {
   return (
     <div className="main-inner-wrapper flex items-stretch gap-2">
       <div className="image">
-        <LoginImage/>
+        <LoginImage />
       </div>
       <div className="form-wrapper w-96 px-[40px] py-[30px] bg-white rounded-md">
         <div className="header">
           <h2>Idea Bank</h2>
-          <h3>Login</h3>
+          <h3>Expert Registration</h3>
         </div>
-        <ExpertLoginForm />
+        <ExpertRegistrationForm />
       </div>
     </div>
   );
 };
 
-export default ExpertLoginFormContent;
+export default ExpertRegistrationFormContent;
