+++ src/ui/Notifications/NotificationBell.jsx
@@ -6,7 +6,7 @@ const NotificationBell = () => {
         <div className="notification-box relative">
             <HiOutlineBellAlert className='text-lg' />
             <div className="badge-count absolute top-[-12px] right-[-5px] rounded-full flex justify-center items-center">
-                5
+                0
             </div>
         </div>
     )
